Moved all image processing into background threads and updated the udpate block on the recognition method to use a CGFloat progress parameter

+++ Overlayer/Utility/SGTextRecognizer.h
@@ -14,6 +14,6 @@
 
 + (SGTextRecognizer *)sharedClient;
 
-- (void)recognizeTextOnImage:(UIImage *)image update:(void (^)(NSUInteger progress))update completion:(void (^)(NSString *recognizedText, NSArray *recognizedCharacterRects))completion;
+- (void)recognizeTextOnImage:(UIImage *)image update:(void (^)(CGFloat progress))update completion:(void (^)(NSString *recognizedText, NSArray *recognizedCharacterRects))completion;
 
 @end
